services/ms-nlp-processor/requirements.txt
python-dotenv
aiohttp
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
langgraph
openai
google-generativeai
anthropic
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
fastapi==0.124.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pipdeptree==2.30.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
rich==13.6.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0